% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert the params.}
\usage{
assert(model, lag)
}
\arguments{
\item{model}{fitted \code{lm(y ~ x)}  object}

\item{lag}{Integer lag length. For example, if the value is \code{10}, results will come for lag lengths of \code{1:10}.}
}
\description{
Assert.
}
\examples{
assert(ts_model,10)
}
