% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-pp.R
\name{summary.pp}
\alias{summary.pp}
\title{PP Test Results.}
\usage{
\method{summary}{pp}(model, lag)
}
\arguments{
\item{model}{fitted \code{lm(y ~ x)}  object}

\item{lag}{Integer lag length. For example, if the value is \code{10}, results will come for lag lengths of \code{1:10}.}
}
\description{
This function computes PP statistics and critical values and aggregate them in one dataframe for all possible situations.
}
\examples{
summary.pp(ts_model,10)
}
