% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-adf.R
\name{summary_adf}
\alias{summary_adf}
\title{ADF Test Results.}
\usage{
summary_adf(model, lag = 5, difference = NA)
}
\arguments{
\item{model}{fitted \code{lm(y ~ x)}  object}

\item{lag}{Integer lag length. For example, if the value is \code{10}, results will come for lag lengths of \code{1:10}.}
}
\description{
This function computes ADF statistics and critical values and aggregate them in one dataframe for all possible situations.
}
\examples{
summary_adf(ts_model,10)
}
