% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-all.R
\name{summary_all}
\alias{summary_all}
\title{Aggregate KPSS, ADF and PP Test Results}
\usage{
summary_all(model, lag, difference = NA)
}
\arguments{
\item{model}{fitted \code{lm(y ~ x)}  object}

\item{lag}{Integer lag length. For example, if the value is \code{10}, results will come for lag lengths of \code{1:10}.}
}
\description{
This function computes unit root test results and aggregate them in one dataframe for all possible situations.
}
\examples{
summary_all(ts_model,10)
}
